Refactor model loading
